impl initState on UseCodeMirrorEditor
