+++ packages/editor/src/stores/codemirror-editor.ts
@@ -1,13 +1,13 @@
 import { useMemo } from 'react';
 
-import { EditorState, type Extension } from '@codemirror/state';
+import { type Extension } from '@codemirror/state';
 import { scrollPastEnd } from '@codemirror/view';
 import {
   type SWRResponseWithUtils, withUtils,
 } from '@growi/core/dist/swr';
 
-import type { UseCodeMirrorEditor, UseCodeMirrorEditorStates } from '../services';
-import { defaultExtensionsToInit, useCodeMirrorEditor } from '../services';
+import type { UseCodeMirrorEditor, UseCodeMirrorEditorResponse } from '../services';
+import { useCodeMirrorEditor } from '../services';
 
 import { useStaticSWR } from './use-static-swr';
 
@@ -15,16 +15,11 @@ const defaultExtensionsMain: Extension[] = [
   scrollPastEnd(),
 ];
 
-const defaultExtensionsToInitMain: Extension[] = [
-  ...defaultExtensionsToInit,
-  ...defaultExtensionsMain,
-];
+type MainEditorUtils = {
+  // impl something
+};
 
-type OperationUtils = {
-  initDoc: (doc?: string) => void,
-}
-
-export const useCodeMirrorEditorMain = (container?: HTMLDivElement | null): SWRResponseWithUtils<OperationUtils, UseCodeMirrorEditorStates> => {
+export const useCodeMirrorEditorMain = (container?: HTMLDivElement | null): SWRResponseWithUtils<MainEditorUtils, UseCodeMirrorEditorResponse> => {
   const props = useMemo<UseCodeMirrorEditor>(() => {
     return {
       container,
@@ -38,19 +33,6 @@ export const useCodeMirrorEditorMain = (container?: HTMLDivElement | null): SWRR
   const swrResponse = useStaticSWR('codeMirrorEditorMain', container != null ? states : undefined);
 
   return withUtils(swrResponse, {
-    initDoc: (doc) => {
-      const currentView = swrResponse.data?.view;
-      if (currentView == null) {
-        return;
-      }
-
-      // create a new state
-      const newState = EditorState.create({
-        doc,
-        extensions: defaultExtensionsToInitMain,
-      });
-
-      currentView.setState(newState);
-    },
+    // impl something
   });
 };
